AuthenticationService.java: Changed login method to find user from UserDB.java and then compare if the password matches the user password.
LoginController.java: Added userSession.login()

ReadinessController.java: Removed auth0 mapping from is-ready to be available at api/is-ready instead of api/auth0/is-ready

UserDB.java:
Added for loop in constructor to simulate database with a lot of users.

Changed the compareUserCredentials method to findUser method, so the method returns user to be compared.

Changed the way method finds the user from for loop to use of hashMap method containsKey() and get().